Update to Ubuntu 20, python 3, and wllvm pkg
This allows us to use newer versions of LLVM and clang. It also uses
python 3, and the corresponding wllvm pip package.

I also decided to use the latest daily R-devel tarball, rather than
the subversion sources. It takes a long time to pull from subversion,
and pulling the tarball is much faster.